Moved unit test to correct directory
